moving training into its own submodule
