+++ deepcell/training/model_trainer.py
@@ -29,42 +29,30 @@ from __future__ import absolute_import
 from __future__ import print_function
 from __future__ import division
 
-import re
 import os
 import json
 import hashlib
 import numpy as np
 
-from deepcell import __version__
 from deepcell.utils.export_utils import export_model
 from deepcell.utils.train_utils import rate_scheduler, get_callbacks
 from deepcell.metrics import Metrics
-
-from tensorflow.keras.optimizers import SGD
 
 
 class ModelTrainer(object):
     def __init__(self,
                  model,
+                 model_name,
+                 model_path,
                  train_generator,
                  validation_generator,
-                 X_train,
-                 y_train,
-                 X_test,
-                 y_test,
-                 model_name = "test_model",
-                 model_path = "test_model_folder",
-                 model_version=0,
                  log_dir=None,
                  tfserving_path=None,
-                 generator_seed=43,
-                 generator_batch_size=1,
-                 generator_transform=None,
-                 generator_transform_kwargs={},
                  training_callbacks='default',
                  postprocessing_fn=None,
                  postprocessing_kwargs={},
                  predict_batch_size=4,
+                 model_version=0,
                  dataset_metadata={},
                  training_kwargs={}):
 
@@ -79,7 +67,7 @@ class ModelTrainer(object):
             model_path (str):
             train_generator (tensorflow.python.keras.preprocessing.image.ImageDataGenerator):
             validation_generator (tensorflow.python.keras.preprocessing.image.ImageDataGenerator):
-            log_dir (str):
+            log_dir (st):
             tfserving_path (str):
             training_callbacks (str):
             postprocessing_fn (function):
@@ -97,27 +85,10 @@ class ModelTrainer(object):
         self.model_version = model_version
 
         # Add dataset information
-        self.X_train = X_train
-        self.y_train = y_train
-        self.X_test = X_test
-        self.y_test = y_test
-
-        # Add generator information
         self.train_generator = train_generator
         self.validation_generator = validation_generator
-        if isinstance(self.model.output_shape, list):
-            self.generator_skips = len(self.model.output_shape) - 1
-        else:
-            self.generator_skips = None
-        self.generator_seed = generator_seed
-        self.generator_batch_size = generator_batch_size
-        self.generator_transform = generator_transform
-        self.generator_transform_kwargs = generator_transform_kwargs
-        print(f'Transform kwargs: {self.generator_transform_kwargs}')
-        # Add miscellaneous information
         self.dataset_metadata = dataset_metadata
         self.postprocessing_fn = postprocessing_fn
-        print(self.postprocessing_fn)
         self.postprocessing_kwargs = postprocessing_kwargs
         self.predict_batch_size = predict_batch_size
 
@@ -136,34 +107,16 @@ class ModelTrainer(object):
         self.batch_size = training_kwargs.pop('batch_size', 1)
         self.training_steps_per_epoch = training_kwargs.pop(
             'training_steps_per_epoch',
-            None)
+            self.train_generator.y.shape[0] // self.batch_size)
         self.validation_steps_per_epoch = training_kwargs.pop(
             'validation_steps_per_epoch',
-            None)
+            self.validation_generator.y.shape[0] // self.batch_size)
         self.n_epochs = training_kwargs.pop('n_epochs', 8)
         self.lr = training_kwargs.pop('lr', 1e-5)
         self.lr_decay = training_kwargs.pop('lr_decay', 0.95)
         self.lr_sched = training_kwargs.pop(
             'lr_sched',
             rate_scheduler(lr=self.lr, decay=self.lr_decay))
-        self.loss_function = training_kwargs.pop("loss_function",
-                "mean_squared_error")
-
-        # create optimizer
-        if "optimizer" in training_kwargs:
-            self.optimizer_passed = True
-            self.optimizer = training_kwargs.pop("optimizer")
-        else:
-            self.optimizer_passed = False
-            self.optimizer_learning_rate = 0.01
-            self.optimizer_decay = 1e-6
-            self.optimizer_momentum = 0.9
-            self.optimizer_nesterov = True
-            self.optimizer = SGD(
-                    lr= self.optimizer_learning_rate,
-                    decay= self.optimizer_decay,
-                    momentum= self.optimizer_momentum,
-                    nesterov= self.optimizer_nesterov)
 
         # Add callbacks
         if training_callbacks == 'default':
@@ -178,282 +131,93 @@ class ModelTrainer(object):
 
         self.trained = False
 
-        self._init_output_metadata()
-
-    def _init_output_metadata(self):
-        """
-        self.output_metadata should ultimately contain human-readable and hashed information
-        on the following components of the training process:
-            deepcell version
-            dataset
-            preprocessing function(s)
-            image generators
-            model (along with initialization state?)
-            postprocessing function(s)
-        """
-        # initialize metadata
-        self.output_metadata = {}
-        # store deepcell version
-        self.output_metadata["deepcell_version"] = __version__
-        # store dataset information
-        self.output_metadata["dataset"] = {}
-        self.output_metadata["dataset"]["training"] = {}
-        self.output_metadata["dataset"]["training"]["X_train_shape"] = self.X_train.shape
-        self.output_metadata["dataset"]["training"]["X_train_datatype"] = self.X_train.dtype.name
-        self.output_metadata["dataset"]["training"]["X_train_md5_digest"] = hashlib.md5(self.X_train).hexdigest()
-        self.output_metadata["dataset"]["training"]["y_train_shape"] = self.y_train.shape
-        self.output_metadata["dataset"]["training"]["y_train_datatype"] = self.y_train.dtype.name
-        self.output_metadata["dataset"]["training"]["y_train_md5_digest"] = hashlib.md5(self.y_train).hexdigest()
-        self.output_metadata["dataset"]["testing"] = {}
-        self.output_metadata["dataset"]["testing"]["X_test_shape"] = self.X_test.shape
-        self.output_metadata["dataset"]["testing"]["X_test_datatype"] = self.X_test.dtype.name
-        self.output_metadata["dataset"]["testing"]["X_test_md5_digest"] = hashlib.md5(self.X_test).hexdigest()
-        self.output_metadata["dataset"]["testing"]["y_test_shape"] = self.y_test.shape
-        self.output_metadata["dataset"]["testing"]["y_test_datatype"] = self.y_test.dtype.name
-        self.output_metadata["dataset"]["testing"]["y_test_md5_digest"] = hashlib.md5(self.y_test).hexdigest()
-
-    def _data_prep(self):
-        
-        def prep_generator(
-                generator_type,
-                output_metadata,
-                data_generator,
-                input_data_dict,
-                skip = None,
-                seed = None,
-                batch_size = None,
-                transform = None,
-                transform_kwargs = {}):
-            """
-            Get training and validation data back from data_generator.flow() and document every step of the process
-            using output_metadata.
-            """
-
-
-            generator_type_string = generator_type + "_generator"
-
-            output_metadata["generators"][generator_type_string] = {}
-            data_generator_name = re.match("<([\w.]+) object at",str(data_generator)).group(1)
-            output_metadata["generators"][generator_type_string]["class"] = data_generator_name
-            
-            parameters = {}
-            parameters["input_data_dict"] = {}
-            for entry in input_data_dict:
-                parameters["input_data_dict"][entry] = {}
-                parameters["input_data_dict"][entry]["md5_digest"] = hashlib.md5(input_data_dict[entry]).hexdigest()
-            parameters["skip"] = skip
-            parameters["seed"] = seed
-            parameters["batch_size"] = batch_size
-            parameters["transform"] = {}
-            try:
-                transform_name = re.match("<([\w.]+) object at",str(transform)).group(1)
-                parameters["transform"]["name"] = transform_name
-                parameters["transform"]["md5_digest"] = hashlib.md5(transform).hexdigest()
-            except AttributeError:
-                pass
-            parameters["transform_kwargs"] = transform_kwargs
-            output_metadata["generators"][generator_type_string]["parameters"] = parameters
-
-            output_data = data_generator.flow(
-                input_data_dict,
-                skip=skip,
-                seed=seed,
-                batch_size=batch_size,
-                transform=transform,
-                transform_kwargs=transform_kwargs)
-
-            return output_data, output_metadata
-
-        self.output_metadata["generators"] = {}
-        ## training image generator
-        train_dict = {"X": self.X_train, "y": self.y_train}
-        self.train_data, self.output_metadata = prep_generator(
-                generator_type = "train",
-                output_metadata = self.output_metadata,
-                data_generator = self.train_generator,
-                input_data_dict = train_dict,
-                skip = self.generator_skips,
-                seed = self.generator_seed,
-                batch_size = self.generator_batch_size,
-                transform = self.generator_transform,
-                transform_kwargs = self.generator_transform_kwargs)
-        ## validation image generator
-        validation_dict = {"X": self.X_test, "y": self.y_test}
-        self.validation_data, self.output_metadata = prep_generator(
-                generator_type = "validation",
-                output_metadata = self.output_metadata,
-                data_generator = self.validation_generator,
-                input_data_dict = validation_dict,
-                skip = self.generator_skips,
-                seed = self.generator_seed,
-                batch_size = self.generator_batch_size,
-                transform = self.generator_transform,
-                transform_kwargs = self.generator_transform_kwargs)
-        print(self.output_metadata)
-
-    def _compile_model(self):
-        
-        def model_compilation(
-                output_metadata,
-                model,
-                loss_function,
-                optimizer,
-                optimizer_flag, 
-                optimizer_learning_rate,
-                optimizer_decay,
-                optimizer_momentum,
-                optimizer_nesterov,
-                metrics):
-            model.compile(loss=loss_function, optimizer=optimizer, metrics=metrics)
-            output_metadata["model"] = {}
-            output_metadata["model"]["compilation"] = {}
-            output_metadata["model"]["compilation"]["loss_function"] = loss_function
-            if optimizer_flag: # optimizer was passed in by user
-                output_metadata["model"]["compilation"]["optimizer"] = "need more information; can't access initialization parameters"
-            else: # optimizer is set to default value
-                output_metadata["model"]["compilation"]["optimizer"] = {}
-                output_metadata["model"]["compilation"]["optimizer"]["class"] = re.match("<([\w.]+) object at",str(optimizer)).group(1)
-                output_metadata["model"]["compilation"]["optimizer"]["learning_rate"] = optimizer_learning_rate
-                output_metadata["model"]["compilation"]["optimizer"]["decay"] = optimizer_decay
-                output_metadata["model"]["compilation"]["optimizer"]["momentum"] = optimizer_momentum
-                output_metadata["model"]["compilation"]["optimizer"]["nesterov"] = optimizer_nesterov
-            output_metadata["model"]["compilation"]["metrics"] = metrics
-
-        metrics = ["accuracy"]
-        model_compilation(
-                self.output_metadata,
-                self.model,
-                self.loss_function,
-                self.optimizer,
-                self.optimizer_passed,
-                self.optimizer_learning_rate,
-                self.optimizer_decay,
-                self.optimizer_momentum,
-                self.optimizer_nesterov,
-                metrics)
-
-    def _train_model(self):
-        if self.training_steps_per_epoch is not None:
-            training_steps_per_epoch = self.training_steps_per_epoch
+    def _create_hash(self):
+        if not self.trained:
+            raise ValueError('Can only create a hash for a trained model')
         else:
-            training_steps_per_epoch = self.train_data.y.shape[0] // self.batch_size
-        
-        if self.validation_steps_per_epoch is not None:
-            validation_steps_per_epoch = self.validation_steps_per_epoch
-        else:
-            validation_steps_per_epoch = self.validation_data.y.shape[0] // self.batch_size
-       
-        def train_model(
-                output_metadata,
-                model,
-                train_data,
-                training_steps_per_epoch,
-                n_epochs,
-                validation_data,
-                validation_steps_per_epoch,
-                training_callbacks):
+            weights = []
+            for layer in self.model.layers:
+                weights += layer.get_weights()
+            summed_weights_list = [np.sum(w) for w in weights]
+            summed_weights = sum(summed_weights_list)
+            model_hash = hashlib.md5(str(summed_weights).encode())
+            self.model_hash = model_hash.hexdigest()
 
-            loss_history = model.fit_generator(
-                train_data,
-                steps_per_epoch=training_steps_per_epoch,
-                epochs=n_epochs,
-                validation_data=validation_data,
-                validation_steps=validation_steps_per_epoch,
-                callbacks=training_callbacks)
-            
-            output_metadata["model"]["training"] = {}
-            output_metadata["model"]["training"]["n_epochs"] = n_epochs
-            output_metadata["model"]["training"]["train_data"] = \
-                    re.match("<([\w.]+) object at",str(train_data)).group(1)
-            output_metadata["model"]["training"]["validation_data"] = \
-                    re.match("<([\w.]+) object at",str(validation_data)).group(1)
-            output_metadata["model"]["training"]["training_steps_per_epoch"] = training_steps_per_epoch
-            output_metadata["model"]["training"]["validation_steps_per_epoch"] = validation_steps_per_epoch
-            output_metadata["model"]["training"]["training_callbacks"] = []
-            for callback in training_callbacks:
-                callback_name = re.match("<([\w.]+) object at",str(callback)).group(1)
-                output_metadata["model"]["training"]["training_callbacks"].append(callback_name)
-
-
-            return loss_history
-
-        def create_hash(trained, model, output_metadata):
-            if not trained:
-                raise ValueError('Can only create a hash for a trained model')
-            else:
-                weights = []
-                for layer in model.layers:
-                    weights += layer.get_weights()
-                summed_weights_list = [np.sum(w) for w in weights]
-                summed_weights = sum(summed_weights_list)
-                model_hash = hashlib.md5(str(summed_weights).encode()).hexdigest()
-                output_metadata["model"]["trained_model_md5_digest"] = model_hash
-                return model_hash
-
-        loss_history = train_model(
-                self.output_metadata,
-                self.model,
-                self.train_data,
-                self.training_steps_per_epoch,
-                self.n_epochs,
-                self.validation_data,
-                self.validation_steps_per_epoch,
-                self.training_callbacks)
+    def _fit(self):
+        loss_history = self.model.fit_generator(
+            self.train_generator,
+            steps_per_epoch=self.training_steps_per_epoch,
+            epochs=self.n_epochs,
+            validation_data=self.validation_generator,
+            validation_steps=self.validation_steps_per_epoch,
+            callbacks=self.training_callbacks)
 
         self.trained = True
         self.loss_history = loss_history
-
-        # Load best performing weights
-        model_name = os.path.join(self.model_path, self.model_name + '.h5')
-        self.model.load_weights(model_name)
-
-        self.model_hash = create_hash(self.trained, self.model, self.output_metadata)
-
-        print(self.output_metadata)
 
     def _benchmark(self):
         if not self.trained:
             raise ValueError('Model training is not complete')
         else:
+            outputs = self.model.predict(
+                self.validation_generator.x,
+                batch_size=self.predict_batch_size)
+            y_pred = self.postprocessing_fn(outputs, **self.postprocessing_kwargs)
+            # TODO: This is a hack because the postprocessing fn returns
+            # masks with no channel dimensions. This should be fixed.
+            y_pred = np.expand_dims(y_pred, axis=-1)
+
+            y_true = self.validation_generator.y.copy()
+            benchmarks = Metrics(self.model_name, seg=False)
+            benchmarks.calc_object_stats(y_true, y_pred)
+
             # Save benchmarks in dict
             self.benchmarks = {}
-            if self.postprocessing_fn:
-                outputs = self.model.predict(
-                    self.validation_data.x,
-                    batch_size=self.predict_batch_size)
-                y_pred = self.postprocessing_fn(outputs, **self.postprocessing_kwargs)
-                # TODO: This is a hack because the postprocessing fn returns
-                # masks with no channel dimensions. This should be fixed.
-                y_pred = np.expand_dims(y_pred, axis=-1)
+            for key in benchmarks.stats.keys():
+                self.benchmarks[key] = int(benchmarks.stats[key].sum())
 
-                y_true = self.validation_data.y.copy()
-                benchmarks = Metrics(self.model_name, seg=False)
-                benchmarks.calc_object_stats(y_true, y_pred)
+    def _create_training_metadata(self):
+        training_metadata = {}
+        training_metadata['batch_size'] = self.batch_size
+        training_metadata['lr'] = self.lr
+        training_metadata['lr_decay'] = self.lr_decay
+        training_metadata['n_epochs'] = self.n_epochs
+        training_metadata['training_steps_per_epoch'] = self.training_steps_per_epoch
+        training_metadata['validation_steps_per_epoch'] = self.validation_steps_per_epoch
 
-                # Save benchmarks
-                for key in benchmarks.stats.keys():
-                    self.benchmarks[key] = int(benchmarks.stats[key].sum())
+        self.training_metadata = training_metadata
 
     def _export_tf_serving(self):
         export_model(self.model, self.tfserving_path, model_version=self.model_version)
 
     def create_model(self, export_serving=False, export_lite=False):
 
-        # Prep data generators
-        self._data_prep()
+        # Train model
+        self._fit()
 
-        # Compile model
-        self._compile_model()
+        # Load best performing weights
+        model_name = os.path.join(self.model_path, self.model_name + '.h5')
+        self.model.load_weights(model_name)
 
-        # Train model with prepped data generators
-        self._train_model()
+        # Create model hash
+        self._create_hash()
 
         # Create benchmarks
         self._benchmark()
 
+        # Create model metadata
+        self._create_training_metadata()
+
         # Save model
         model_name = os.path.join(self.model_path, self.model_name + '_' + self.model_hash + '.h5')
         self.model.save(model_name)
+
+        # Save metadata (training and dataset) and benchmarks
+        metadata = {}
+        metadata['model_hash'] = self.model_hash
+        metadata['training_metadata'] = self.training_metadata
+        metadata['dataset_metadata'] = self.dataset_metadata
+        metadata['benchmarks'] = self.benchmarks
 
         # TODO: Saving the benchmarking object in this way saves each individual benchmark.
         # This should be refactored to save the sums.
@@ -462,11 +226,8 @@ class ModelTrainer(object):
             self.model_name, self.model_hash))
 
         with open(metadata_name, 'w') as json_file:
-            json.dump(self.output_metadata, json_file)
+            json.dump(metadata, json_file)
 
         # Export tf serving model
         if export_serving:
             self._export_tf_serving()
-
-        # return information to calling program
-        return model_name, metadata_name, self.model_hash
